change GameRunner to take PlayerSpecs by value instead of by ref to reduce risk of unexpected Pybind behavior
