+++ src/xiangqipy/command_input.py
@@ -13,69 +13,7 @@ import xiangqi_bindings as xb
 from xiangqipy.enums import EvaluatorType, PlayerType
 
 
-@dataclass
-class RawPlayerInput:
-    color: xb.PieceColor
-    player_type: Literal[None, "ai", "person"]
-    algo: Literal[None, "minmax", "random"]
-    strength: int = None
-    key_size: int = None
-    number_zobrist_states: int = None
-    zkeys_seed: int = None
 
-    @property
-    def evaluator_type(self) -> xb.EvaluatorType:
-        if self.player_type in [None, "ai"]:
-            if self.algo in ["minimax", None]:
-                return xb.EvaluatorType.kMinimax
-            if self.algo == "random":
-                return xb.EvaluatorType.kRandom
-        else:
-            return xb.EvaluatorType.kHuman
-
-    @property
-    def minimax_search_depth(self) -> int:
-        if self.evaluator_type in [
-            xb.EvaluatorType.kRandom,
-            xb.EvaluatorType.kHuman,
-        ]:
-            return 0
-        elif self.evaluator_type == xb.EvaluatorType.kMinimax:
-            if self.strength is None:
-                return 4
-            else:
-                return self.strength
-
-    @property
-    def zobrist_key_size_bits(self) -> int:
-        if self.evaluator_type in [
-            xb.EvaluatorType.kRandom,
-            xb.EvaluatorType.kHuman,
-        ]:
-            return 0
-        elif self.evaluator_type == xb.EvaluatorType.kMinimax:
-            return self.key_size if self.key_size else 64
-
-    @property
-    def zobrist_calculator_count(self) -> int:
-        if self.evaluator_type in [
-            xb.EvaluatorType.kRandom,
-            xb.EvaluatorType.kHuman,
-        ]:
-            return 0
-        elif self.evaluator_type == xb.EvaluatorType.kMinimax:
-            return (
-                self.number_zobrist_states if self.number_zobrist_states else 2
-            )
-
-    def to_player_spec(self) -> xb.PlayerSpec:
-        return xb.PlayerSpec(
-            color=self.color,
-            evaluator_type=self.evaluator_type,
-            zobrist_key_size_bits=self.zobrist_key_size_bits,
-            zobrist_calculator_count=self.zobrist_calculator_count,
-            minimax_search_depth=self.minimax_search_depth,
-        )
 
 
 @dataclass
